PhotoBlog -  Small Project
This is a basic PhotoBlog website with quite a nice minimalistic design. I used pictures from Unsplash, a great resource to get free pictures to use. 
The main aim of this short project was to get used to properly used google-fonts and become familiar with using padding and margin effectively.