Pricing Panel - Small Project
This is a step up from the last project. Did some more work utilizing margins and padding. Introduced some new concepts such as justify-content, align-items, and media queries.  Using media queries was new for me, it took a bit of time to get used to and understand the concept but I feel that now I have a good understanding of what it is used for, and looking forward to practicing and utilising it much more in the future