The Black Mamba - Medium Project
Largest Project up to date - first project that used more than 150 lines of VSCode. This was definitely a step up in terms of what I had previously been doing. Worked with Bootstrap CSS during this project. Utilized columns, rows, and IMG-fluid to create a fluid, responsive webpage. This was based on the 'Museum of Candy' project on the course.  

Noted - I do not own the rights to the Black Mamba Logo -  taken from https://1000logos.net/wp-content/uploads/2017/10/Kobe-bryant-Logo.png